# deployment/perfulandia_documentacion_openapi.yaml
openapi: 3.0.1
info:
  title: Perfulandia
  description: Documentación OPENAPI de perfulandia
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /support/respond/{ticketId}:
    put:
      tags:
      - support-controller
      operationId: respondTicket
      parameters:
      - name: ticketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support/close/{ticketId}:
    put:
      tags:
      - support-controller
      operationId: closeTicket
      parameters:
      - name: ticketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /ship/user/{userId}:
    put:
      tags:
      - shipping-controller
      operationId: shipOrder
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
  /product:
    get:
      tags:
      - product-controller
      operationId: getAllProducts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    put:
      tags:
      - product-controller
      operationId: updateProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
    post:
      tags:
      - product-controller
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
  /product/{productId}/inventory/{quantity}:
    put:
      tags:
      - product-controller
      operationId: assignInventory
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: quantity
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
  /product/{productId}/category/{categoryId}:
    put:
      tags:
      - product-controller
      operationId: assignCategory
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: categoryId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
  /employees:
    get:
      tags:
      - employee-controller
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    put:
      tags:
      - employee-controller
      operationId: update
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Employee"
    post:
      tags:
      - employee-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}/role/{roleId}:
    put:
      tags:
      - employee-controller
      operationId: assignRole
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: roleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Employee"
  /cart/{userId}/quantity/{itemId}:
    put:
      tags:
      - cart-controller
      operationId: updateItemQuantity
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: quantity
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Cart"
  /user:
    get:
      tags:
      - user-controller
      operationId: getAllUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags:
      - user-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /support/new/{userId}:
    post:
      tags:
      - support-controller
      operationId: newSupport
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicket"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /order/pay/{userId}:
    post:
      tags:
      - order-controller
      operationId: payPendingOrder
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
  /order/checkout/{userId}:
    post:
      tags:
      - order-controller
      operationId: createOrder
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
  /cart/{userId}:
    get:
      tags:
      - cart-controller
      operationId: getCartByUserId
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      tags:
      - cart-controller
      operationId: createCart
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{userId}/add:
    post:
      tags:
      - cart-controller
      operationId: addToCart
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItem"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Cart"
  /user/{id}:
    get:
      tags:
      - user-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
  /support:
    get:
      tags:
      - support-controller
      operationId: getAllTickets
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
  /support/user/{userId}:
    get:
      tags:
      - support-controller
      operationId: getTicket
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
  /support/user/{userId}/open:
    get:
      tags:
      - support-controller
      operationId: getOpenTicket
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
  /order/{userId}:
    get:
      tags:
      - order-controller
      operationId: getOrders
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /order/pending/{userId}:
    get:
      tags:
      - order-controller
      operationId: getPendingOrders
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /employees/{id}:
    get:
      tags:
      - employee-controller
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      tags:
      - employee-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /product/{id}:
    delete:
      tags:
      - product-controller
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    Cart:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/components/schemas/User"
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
    CartItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cart:
          $ref: "#/components/schemas/Cart"
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
          format: int32
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderProducts:
          type: array
          items:
            $ref: "#/components/schemas/OrderProduct"
        orderDate:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        total:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - PAYMENT_PENDING
          - PROCESSING
          - SHIPPED
          - COMPLETED
          - CANCELLED
        paymentDate:
          type: string
          format: date-time
        totalPaid:
          type: integer
          format: int32
    OrderProduct:
      type: object
      properties:
        id:
          type: integer
          format: int64
        order:
          $ref: "#/components/schemas/Order"
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
          format: int32
        buyPrice:
          type: integer
          format: int64
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        brand:
          type: string
        notes:
          type: string
        price:
          type: integer
          format: int32
        category:
          $ref: "#/components/schemas/Category"
        ml:
          type: integer
          format: int32
        inventory:
          type: integer
          format: int32
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
          format: int64
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - CLOSED
        fechaCreacion:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        answer:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        password:
          type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        cart:
          $ref: "#/components/schemas/Cart"
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        sucursal:
          type: string
        activo:
          type: boolean
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string